AutoDiff

Abstract Types

ExaPF.AutoDiff.AbstractHessianProd — Type

AbstractHessianProd

Returns the adjoint-Hessian-vector product $λ^⊤ H v$ of any nonlinear constraint $h(x)$.

ExaPF.AutoDiff.AbstractFullHessian — Type

AbstractHessianProd

Full sparse Hessian $H$ of any nonlinear constraint $h(x)$.

ExaPF.AutoDiff.AbstractJacobian — Type

AbstractJacobian

Automatic differentiation for the compressed Jacobian of any nonlinear constraint $h(x)$.

ExaPF.AutoDiff.AbstractStack — Type

AbstractStack{VT}

Abstract variable storing the inputs and the intermediate values in the expression tree.

AutoDiff API

ExaPF.AutoDiff.seed! — Function

seed!(
    H::AbstractHessianProd,
    v::AbstractVector{T},
) where {T}

Seed the duals with v to compute the Hessian vector product $λ^⊤ H v$.

ExaPF.AutoDiff.seed_coloring! — Function

seed_coloring!(
    M::Union{AbstractJacobian, AbstractFullHessian}
    coloring::AbstractVector,
)

Seed the duals with the coloring based seeds to compute the Jacobian or Hessian $M$.

ExaPF.AutoDiff.partials! — Function

partials!(jac::AbstractJacobian)

Extract partials from Jacobian jac in jac.J.

partials!(hess::AbstractFullHessian)

Extract partials from Hessian hess into hess.H.

ExaPF.AutoDiff.set_value! — Function

set_value!(
    jac,
    primals::AbstractVector{T}
) where {T}

Set values of ForwardDiff.Dual numbers in jac to primals.