Polar formulation

Generic templates

ExaPF.AbstractVariable — Type

AbstractVariable

Variables corresponding to a particular formulation.

ExaPF.AbstractFormulation — Type

AbstractFormulation

Interface between the data and the mathemical formulation.

ExaPF.State — Type

State <: AbstractVariable

All variables $x$ depending on the variables Control $u$ through the non-linear equation $g(x, u) = 0$.

ExaPF.Control — Type

Control <: AbstractVariable

Independent variables $u$ used in the reduced-space formulation.

Structure and variables

ExaPF.PolarForm — Type

PolarForm{T, IT, VT, MT} <: AbstractFormulation

Wrap a PS.PowerNetwork network to load the data on the target device (CPU() and CUDADevice() are currently supported).

ExaPF.NetworkStack — Type

NetworkStack <: AbstractStack
NetworkStack(polar::PolarForm)
NetworkStack(nbus::Int, ngen::Int, nlines::Int, VT::Type)

Store the variables associated to the polar formulation. The variables are stored in the field input, ordered as follows

    input = [vmag ; vang ; pgen]

The object stores also intermediate variables needed in the expression tree, such as the LKMR basis ψ.

Notes

The NetworkStack can be instantiated on the host or on the target device.

ExaPF.init! — Function

init!(polar::PolarForm, stack::NetworkStack)

Set stack.input with the initial values specified in the base PS.PowerNetwork object.

The state and the control are defined as mapping:

ExaPF.mapping — Function

mapping(polar::PolarForm, ::State)

Return the mapping associated to the State() in NetworkStack according to the polar formulation PolarForm.

mapping(polar::PolarForm, ::Control)

Return the mapping associated to the Control() in NetworkStack according to the polar formulation PolarForm.

Powerflow solver

ExaPF.run_pf — Function

run_pf(
    polar::PolarForm, stack::NetworkStack;
    rtol=1e-8, max_iter=20, verbose=0,
)

Solve the power flow equations $g(x, u) = 0$ w.r.t. the stack $x$, using the (NewtonRaphson algorithm. The initial state $x$ is specified implicitly inside stack, with the mapping mapping associated to the polar formulation. The object stack is modified inplace in the function.

The algorithm stops when a tolerance rtol or a maximum number of iterations maxiter is reached.

Arguments

polar::AbstractFormulation: formulation of the power flow equation

stack::NetworkStack: initial values in the network

ExaPF.NewtonRaphson — Type

NewtonRaphson <: AbstractNonLinearSolver

Newton-Raphson algorithm.

Attributes

maxiter::Int (default 20): maximum number of iterations

tol::Float64 (default 1e-8): tolerance of the algorithm

verbose::Int (default 0): verbosity level

Constraints

The different parts of the polar formulation are implemented in the following AbstractExpression:

ExaPF.PolarBasis — Type

PolarBasis{VI, MT} <: AbstractExpression
PolarBasis(polar::PolarForm)

Implement the LKMR nonlinear basis. Takes as input the voltage magnitudes vmag and the voltage angles vang and returns

\[ \begin{aligned} & \psi_\ell^C(v, \theta) = v^f v^t \cos(\theta_f - \theta_t) \quad \forall \ell = 1, \cdots, n_\ell \\ & \psi_\ell^S(v, \theta) = v^f v^t \sin(\theta_f - \theta_t) \quad \forall \ell = 1, \cdots, n_\ell \\ & \psi_k(v, \theta) = v_k^2 \quad \forall k = 1, \cdots, n_b \end{aligned}\]

Dimension: 2 * n_lines + n_bus

Complexity

3 n_lines + n_bus mul, n_lines cos and n_lines sin

ExaPF.PowerFlowBalance — Type

PowerFlowBalance{VT, MT}
PowerFlowBalance(polar)

Implement a subset of the power injection corresponding to $(p_{inj}^{pv}, p_{inj}^{pq}, q_{inj}^{pq})$. The function encodes the active balance equations at PV and PQ nodes, and the reactive balance equations at PQ nodes:

\[\begin{aligned} p_i &= v_i \sum_{j}^{n} v_j (g_{ij}\cos{(\theta_i - \theta_j)} + b_{ij}\sin{(\theta_i - \theta_j})) \,, & ∀ i ∈ \{PV, PQ\} \\ q_i &= v_i \sum_{j}^{n} v_j (g_{ij}\sin{(\theta_i - \theta_j)} - b_{ij}\cos{(\theta_i - \theta_j})) \,. & ∀ i ∈ \{PQ\} \end{aligned}\]

Dimension: n_{pv} + 2 * n_{pq}

Complexity

2 SpMV

ExaPF.VoltageMagnitudeBounds — Type

VoltageMagnitudeBounds

Implement the bounds on voltage magnitudes not taken into account in the bound constraints. In the reduced space, this is associated to the the voltage magnitudes at PQ nodes:

\[v_{pq}^♭ ≤ v_{pq} ≤ v_{pq}^♯ .\]

Dimension: n_pq

Complexity

1 copyto

Note

In the reduced space, the constraints on the voltage magnitudes at PV nodes $v_{pv}$ are taken into account when bounding the control $u$.

ExaPF.PowerGenerationBounds — Type

PowerGenerationBounds{VT, MT}
PowerGenerationBounds(polar)

Constraints on the active power productions and on the reactive power productions that are not already taken into account in the bound constraints. In the reduced space, that amounts to

\[p_{g,ref}^♭ ≤ p_{g,ref} ≤ p_{g,ref}^♯ ; C_g q_g^♭ ≤ C_g q_g ≤ C_g q_g^♯ .\]

Dimension: n_pv + 2 n_ref

Complexity

1 copyto, 1 SpMV

ExaPF.LineFlows — Type

LineFlows{VT, MT}
LineFlows(polar)

Implement thermal limit constraints on the lines of the network.

Dimension: 2 * n_lines

Complexity

4 SpMV, 4 * n_lines quadratic, 2 * n_lines add

Objective

The production costs is given in the AbstractExpression CostFunction:

ExaPF.CostFunction — Type

CostFunction{VT, MT} <: AutoDiff.AbstractExpression
CostFunction(polar)

Implement the quadratic cost function for OPF

\[ ∑_{g=1}^{n_g} c_{2,g} p_g^2 + c_{1,g} p_g + c_{0,g}\]

Dimension: 1

Complexity

1 SpMV, 1 sum

Composition of expressions

The different expressions can be combined together in several different ways.

ExaPF.MultiExpressions — Type

MultiExpressions <: AbstractExpression

Implement expressions concatenation. Takes as input a vector of expressions [expr1,...,exprN] and concatenate them in a single expression mexpr, such that

    mexpr(x) = [expr1(x) ; expr2(x) ; ... ; exprN(x)]

ExaPF.ComposedExpressions — Type

ComposedExpressions{Expr1<:PolarBasis, Expr2} <: AbstractExpression

Implement expression composition. Takes as input two expressions expr1 and expr2 and returns a composed expression cexpr such that ``` cexpr(x) = expr2 ∘ expr1(x)

Notes

Currently, only PolarBasis is supported for expr1.