Formulations

Description

ExaPF.AbstractFormulation — Type

AbstractFormulation

Second layer of the package, implementing the interface between the first layer (the topology of the network) and the third layer (implementing the callbacks for the optimization solver).

ExaPF.PolarForm — Type

PolarForm{T, IT, VT, MT}

Wrap a PS.PowerNetwork network to load the data on the target device (CPU() and CUDADevice() are currently supported).

Powerflow solver

Missing docstring.

Missing docstring for powerflow. Check Documenter's build log for details.

ExaPF.NewtonRaphson — Type

NewtonRaphson <: AbstractNonLinearSolver

Newton-Raphson algorithm.

Attributes

maxiter::Int (default 20): maximum number of iterations

tol::Float64 (default 1e-8): tolerance of the algorithm

verbose::Int (default 0): verbosity level

Constraints

Current supported constraints are:

Missing docstring.

Missing docstring for voltage_magnitude_constraints. Check Documenter's build log for details.

Missing docstring.

Missing docstring for active_power_constraints. Check Documenter's build log for details.

Missing docstring.

Missing docstring for reactive_power_constraints. Check Documenter's build log for details.

Missing docstring.

Missing docstring for flow_constraints. Check Documenter's build log for details.

Missing docstring.

Missing docstring for power_balance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bus_power_injection. Check Documenter's build log for details.

Missing docstring.

Missing docstring for network_operations. Check Documenter's build log for details.

These functions allow to query constraints' attributes:

Missing docstring.

Missing docstring for is_constraint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for size_constraint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bounds. Check Documenter's build log for details.

ExaPF implements special functions to compute the derivatives of each constraints:

Missing docstring.

Missing docstring for adjoint!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jacobian_transpose_product!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for matpower_jacobian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for matpower_hessian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jacobian_sparsity. Check Documenter's build log for details.

API Reference

Variables

ExaPF.AbstractVariable — Type

AbstractVariable

Variables corresponding to a particular formulation.

ExaPF.State — Type

State <: AbstractVariable

All variables $x$ depending on the variables Control $u$ through the non-linear equation $g(x, u) = 0$.

ExaPF.Control — Type

Control <: AbstractVariable

Independent variables $u$ used in the reduced-space formulation.

Missing docstring.

Missing docstring for PhysicalState. Check Documenter's build log for details.

Get default values attached to a given variable:

Missing docstring.

Missing docstring for initial. Check Documenter's build log for details.

Attributes

Missing docstring.

Missing docstring for AbstractFormAttribute. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NumberOfState. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NumberOfControl. Check Documenter's build log for details.

ExaPF extends Base.get to query the different attributes of a model:

Base.get — Function

get(pf::AbstractPowerSystem, attr::AbstractNetworkAttribute)

Return value of attribute attr in the AbstractPowerSystem object pf.

get(pf::AbstractPowerSystem, attr::AbstractIndexing)

Return indexing corresponding to a subset of the buses.

Examples

npq = get(pf, NumberOfPQBuses())
npv = get(pf, NumberOfPVBuses())

The associated setter is implemented with setvalues!:

Missing docstring.

Missing docstring for setvalues!. Check Documenter's build log for details.

Costs

Missing docstring.

Missing docstring for cost_production. Check Documenter's build log for details.