Linear solvers

Description

ExaPF allows to solve linear systems with either direct and indirect linear algebra, both on CPU and on GPU. To solve a linear system $Ax = b$, ExaPF uses the function ldiv!.

ExaPF.LinearSolvers.ldiv! — Function

ldiv!(solver, x, A, y)
ldiv!(solver, x, y)

solver::AbstractLinearSolver: linear solver to solve the system

x::AbstractVector: Solution

A::AbstractMatrix: Input matrix

y::AbstractVector: RHS

Solve the linear system $A x = y$ using the algorithm specified in solver. If A is not specified, the function will used directly the factorization stored inside solver.

Direct solvers

ExaPF wraps UMFPACK (shipped with SuiteSparse.jl) on the CPU, and CUSPARSE on CUDA device.

ExaPF.LinearSolvers.DirectSolver — Type

DirectSolver <: AbstractLinearSolver

Solve linear system $A x = y$ with direct linear algebra.

On the CPU, DirectSolver uses UMFPACK to solve the linear system

On CUDA GPU, DirectSolver redirects the resolution to the method CUSOLVER.csrlsvqr

Iterative solvers

ExaPF.LinearSolvers.Bicgstab — Type

Bicgstab <: AbstractIterativeLinearSolver
Bicgstab(precond; verbose=0, rtol=1e-10, atol=1e-10)

Wrap Krylov.jl's BICGSTAB algorithm to solve iteratively the linear system $A x = y$.

ExaPF.LinearSolvers.Dqgmres — Type

Dqgmres <: AbstractIterativeLinearSolver
Dqgmres(precond; verbose=false, memory=4)

Wrap Krylov.jl's Dqgmres algorithm to solve iteratively the linear system $A x = y$.

Available linear solvers can be queried with

ExaPF.LinearSolvers.list_solvers — Function

list_solvers(::KernelAbstractions.Device)

List linear solvers available on current device.

A default solver is provided for each vendor backend.

ExaPF.LinearSolvers.default_linear_solver — Function

default_linear_solver(A, ::KA.CPU)

Default linear solver on the CPU.

Block-Krylov solvers

ExaPF.jl provides an implementation of block-GMRES to solve linear systems with multiple righ-hand sides.

ExaPF.LinearSolvers.BlockKrylovSolver — Type

Abstract type for using block Krylov solvers in-place

ExaPF.LinearSolvers.BlockGmresSolver — Type

Type for storing the vectors required by the in-place version of BLOCK-GMRES.

The outer constructors

solver = BlockGmresSolver(m, n, p, memory, SV, SM)
solver = BlockGmresSolver(A, B; memory=5)

may be used in order to create these vectors. memory is set to div(n,p) if the value given is larger than div(n,p).

ExaPF.LinearSolvers.block_gmres — Function

(X, stats) = block_gmres(A, B; X0::AbstractMatrix{FC}; memory::Int=20, M=I, N=I,
                         ldiv::Bool=false, restart::Bool=false, reorthogonalization::Bool=false,
                         atol::T = √eps(T), rtol::T=√eps(T), itmax::Int=0,
                         timemax::Float64=Inf, verbose::Int=0, history::Bool=false)

T is an AbstractFloat such as Float32, Float64 or BigFloat. FC is T or Complex{T}.

(X, stats) = block_gmres(A, B, X0::AbstractVector; kwargs...)

GMRES can be warm-started from an initial guess X0 where kwargs are the same keyword arguments as above.

Solve the linear system AX = B of size n with p right-hand sides using block-GMRES.

Input arguments

A: a linear operator that models a matrix of dimension n;

B: a matrix of size n × p.

Optional argument

X0: a matrix of size n × p that represents an initial guess of the solution X.

Keyword arguments

memory: if restart = true, the restarted version block-GMRES(k) is used with k = memory. If restart = false, the parameter memory should be used as a hint of the number of iterations to limit dynamic memory allocations. Additional storage will be allocated if the number of iterations exceeds memory;

M: linear operator that models a nonsingular matrix of size n used for left preconditioning;

N: linear operator that models a nonsingular matrix of size n used for right preconditioning;

ldiv: define whether the preconditioners use ldiv! or mul!;

restart: restart the method after memory iterations;

reorthogonalization: reorthogonalize the new matrices of the block-Krylov basis against all previous matrix;

atol: absolute stopping tolerance based on the residual norm;

rtol: relative stopping tolerance based on the residual norm;

itmax: the maximum number of iterations. If itmax=0, the default number of iterations is set to 2 * div(n,p);

timemax: the time limit in seconds;

verbose: additional details can be displayed if verbose mode is enabled (verbose > 0). Information will be displayed every verbose iterations;

history: collect additional statistics on the run such as residual norms.

Output arguments

x: a dense matrix of size n × p;

stats: statistics collected on the run in a BlockGmresStats.

ExaPF.LinearSolvers.block_gmres! — Function

solver = block_gmres!(solver::BlockGmresSolver, B; kwargs...)
solver = block_gmres!(solver::BlockGmresSolver, B, X0; kwargs...)

where kwargs are keyword arguments of block_gmres.

See BlockGmresSolver for more details about the solver.