Linear solvers

Description

ExaPF allows to solve linear systems with either direct and indirect linear algebra, both on CPU and on GPU. To solve a linear system $Ax = b$, ExaPF uses the function ldiv!.

LinearAlgebra.ldiv! — Function

ldiv!(solver, x, A, y)
ldiv!(solver, x, y)

solver::AbstractLinearSolver: linear solver to solve the system

x::AbstractVector: Solution

A::AbstractMatrix: Input matrix

y::AbstractVector: RHS

Solve the linear system $A x = y$ using the algorithm specified in solver. If A is not specified, the function will used directly the factorization stored inside solver.

Direct solvers

ExaPF wraps UMFPACK (shipped with SuiteSparse.jl) on the CPU, and CUSPARSE on CUDA device.

ExaPF.LinearSolvers.DirectSolver — Type

DirectSolver <: AbstractLinearSolver

Solve linear system $A x = y$ with direct linear algebra.

On the CPU, DirectSolver uses UMFPACK to solve the linear system

On CUDA GPU, DirectSolver redirects the resolution to the method CUSOLVER.csrlsvqr

Iterative solvers

ExaPF.LinearSolvers.KrylovBICGSTAB — Type

KrylovBICGSTAB <: AbstractIterativeLinearSolver
KrylovBICGSTAB(precond; verbose=0, rtol=1e-10, atol=1e-10)

Wrap Krylov.jl's BICGSTAB algorithm to solve iteratively the linear system $A x = y$.

ExaPF.LinearSolvers.DQGMRES — Type

DQGMRES <: AbstractIterativeLinearSolver
DQGMRES(precond; verbose=false, memory=4)

Wrap Krylov.jl's DQGMRES algorithm to solve iteratively the linear system $A x = y$.

ExaPF.LinearSolvers.BICGSTAB — Type

BICGSTAB <: AbstractIterativeLinearSolver
BICGSTAB(precond; maxiter=2_000, tol=1e-8, verbose=false)

Custom BICGSTAB implementation to solve iteratively the linear system $A x = y$.

ExaPF.LinearSolvers.EigenBICGSTAB — Type

EigenBICGSTAB <: AbstractIterativeLinearSolver
EigenBICGSTAB(precond; maxiter=2_000, tol=1e-8, verbose=false)

Julia's port of Eigen's BICGSTAB to solve iteratively the linear system $A x = y$.

ExaPF.jl is shipped with a custom BICGSTAB implementation. However, we highly recommend to use KrylovBICGSTAB instead, which has proved to be more robust.

ExaPF.LinearSolvers.bicgstab — Function

bicgstab(A, b, P, xi;
         tol=1e-8,
         maxiter=size(A, 1),
         verbose=false,
         maxtol=1e20)

BiCGSTAB implementation according to

Van der Vorst, Henk A. "Bi-CGSTAB: A fast and smoothly converging variant of Bi-CG for the solution of nonsymmetric linear systems." SIAM Journal on scientific and Statistical Computing 13, no. 2 (1992): 631-644.

Available linear solvers can be queried with

ExaPF.LinearSolvers.list_solvers — Function

list_solvers(::KernelAbstractions.Device)

List linear solvers available on current device.

Preconditioning

To solve linear systems with iterative methods, ExaPF provides an implementation of a block-Jacobi preconditioner, portable on GPU.

ExaPF.LinearSolvers.AbstractPreconditioner — Type

AbstractPreconditioner

Preconditioners for the iterative solvers mostly focused on GPUs

Block-Jacobi preconditioner

ExaPF.LinearSolvers.BlockJacobiPreconditioner — Type

BlockJacobiPreconditioner

Creates an object for the block-Jacobi preconditioner

nblocks::Int64: Number of partitions or blocks.

blocksize::Int64: Size of each block.

nJs::Int64: Size of the blocks. For the GPUs these all have to be of equal size.

partitions::Vector{Vector{Int64}}:npart` partitions stored as lists

cupartitions: partitions transfered to the GPU

lpartitions::Vector{Int64}`: Length of each partitions.

culpartitions::Vector{Int64}`: Length of each partitions, on the GPU.

blocks: Dense blocks of the block-Jacobi

cublocks: Js transfered to the GPU

map: The partitions as a mapping to construct views

cumap: cumap transferred to the GPU`

part: Partitioning as output by Metis

cupart: part transferred to the GPU

P: The sparse precondition matrix whose values are updated at each iteration

ExaPF.LinearSolvers.update — Function

function update(J::CuSparseMatrixCSR, p)

Update the preconditioner p from the sparse Jacobian J in CSR format for the GPU

The dense blocks cuJs are filled from the sparse Jacobian J

To a batch inversion of the dense blocks using CUBLAS

Extract the preconditioner matrix p.P from the dense blocks cuJs

function update(J::SparseMatrixCSC, p)

Update the preconditioner p from the sparse Jacobian J in CSC format for the CPU

Note that this implements the same algorithm as for the GPU and becomes very slow on CPU with growing number of blocks.

ExaPF.LinearSolvers.build_adjmatrix — Function

build_adjmatrix

Build the adjacency matrix of a matrix A corresponding to the undirected graph

ExaPF.LinearSolvers.fillblock_gpu! — Function

fillblock_gpu

Fill the dense blocks of the preconditioner from the sparse CSR matrix arrays

ExaPF.LinearSolvers.fillP_gpu! — Function

fillP_gpu

Update the values of the preconditioner matrix from the dense Jacobi blocks