PowerSystem

Description

ExaPF.PowerSystem.AbstractPowerSystem — Type

AbstractPowerSystem

First layer of the package. Store the topology of a given transmission network, including:

the power injection at each bus ;

the admittance matrix ;

the default voltage at each bus.

Data are imported either from a matpower file, or a PSSE file.

ExaPF.PowerSystem.PowerNetwork — Type

PowerNetwork <: AbstractPowerSystem

This structure contains constant parameters that define the topology and physics of the power network.

The object PowerNetwork uses its own contiguous indexing for the buses. The indexing is independent from those specified in the Matpower or the PSSE input file. However, a correspondence between the two indexing (Input indexing to PowerNetwork indexing) is stored inside the attribute bus_to_indexes.

Note

The object PowerNetwork is created in the host memory. Use a AbstractFormulation object to move data to the target backend.

ExaPF.PowerSystem.load_case — Function

load_case(case_name, lib::PowerNetworkLibrary=EXADATA)

Convenient function to load a PowerNetwork instance from one of the benchmark library (dir=EXADATA for MATPOWER, dir=PGLIB for PGLIB-OPF). Default library is lib=EXADATA.

Examples

julia> net = PS.load_case("case118") # default is MATPOWER
PowerNetwork object with:
    Buses: 118 (Slack: 1. PV: 53. PQ: 64)
    Generators: 54.

julia> net = PS.load_case("case1354_pegase", PS.PGLIB)
PowerNetwork object with:
    Buses: 1354 (Slack: 1. PV: 259. PQ: 1094)
    Generators: 260.

API Reference

Network elements

ExaPF.PowerSystem.AbstractNetworkElement — Type

AbstractNetworkElement

Abstraction for all physical elements being parts of a AbstractPowerSystem. Elements are divided in

transmission lines (Lines)

buses (Buses)

generators (Generators)

List of elements:

ExaPF.PowerSystem.Buses — Type

Buses <: AbstractNetworkElement

Buses of a transmission network.

ExaPF.PowerSystem.Lines — Type

Lines <: AbstractNetworkElement

Lines of a transmission network.

ExaPF.PowerSystem.Generators — Type

Generators <: AbstractElement

Generators in a transmission network

Network attributes

ExaPF.PowerSystem.AbstractNetworkAttribute — Type

AbstractNetworkAttribute

Attribute of a AbstractPowerSystem.

List of attributes:

ExaPF.PowerSystem.NumberOfBuses — Type

NumberOfBuses <: AbstractNetworkAttribute

Number of buses in a AbstractPowerSystem.

ExaPF.PowerSystem.NumberOfLines — Type

NumberOfLines <: AbstractNetworkAttribute

Number of lines in a AbstractPowerSystem.

ExaPF.PowerSystem.NumberOfGenerators — Type

NumberOfGenerators <: AbstractNetworkAttribute

Number of generators in a AbstractPowerSystem.

ExaPF.PowerSystem.NumberOfPVBuses — Type

NumberOfPVBuses <: AbstractNetworkAttribute

Number of PV buses in a AbstractPowerSystem.

ExaPF.PowerSystem.NumberOfPQBuses — Type

NumberOfPQBuses <: AbstractNetworkAttribute

Number of PQ buses in a AbstractPowerSystem.

ExaPF.PowerSystem.NumberOfSlackBuses — Type

NumberOfSlackBuses <: AbstractNetworkAttribute

Number of slack buses in a AbstractPowerSystem.

ExaPF.PowerSystem.BaseMVA — Type

BaseMVA <: AbstractNetworkAttribute

Base MVA of the network.

ExaPF.PowerSystem.BusAdmittanceMatrix — Type

BusAdmittanceMatrix <: AbstractNetworkAttribute

Bus admittance matrix associated with the topology of the network.

Query the indexing of the different elements in a given network:

ExaPF.PowerSystem.PVIndexes — Type

PVIndexes <: AbstractIndexing

Indexes of the PV buses in a AbstractPowerSystem.

ExaPF.PowerSystem.PQIndexes — Type

PQIndexes <: AbstractIndexing

Indexes of the PQ buses in a AbstractPowerSystem.

ExaPF.PowerSystem.SlackIndexes — Type

SlackIndexes <: AbstractIndexing

Indexes of the slack buses in a AbstractPowerSystem.

ExaPF.PowerSystem.GeneratorIndexes — Type

GeneratorIndexes <: AbstractIndexing

Indexes of the generators in a AbstractPowerSystem.

Network values

ExaPF.PowerSystem.AbstractNetworkValues — Type

AbstractNetworkValues

Numerical values attached to the different attributes of the network.

List of values:

ExaPF.PowerSystem.VoltageMagnitude — Type

VoltageMagnitude <: AbstractNetworkValues

Magnitude |v| of the voltage v = |v| exp(i θ).

ExaPF.PowerSystem.VoltageAngle — Type

VoltageAngle <: AbstractNetworkValues

Angle θ of the voltage v = |v| exp(i θ).

ExaPF.PowerSystem.ActivePower — Type

ActivePower <: AbstractNetworkValues

Active power P of the complex power S = P + iQ.

ExaPF.PowerSystem.ReactivePower — Type

ReactivePower <: AbstractNetworkValues

Reactive power Q of the complex power S = P + iQ.

ExaPF.PowerSystem.ActiveLoad — Type

ActiveLoad <: AbstractNetworkValues

Active load Pd at buses.

ExaPF.PowerSystem.ReactiveLoad — Type

ReactiveLoad <: AbstractNetworkValues

Reactive load Qd at buses.

Function to get the range of a given value:

ExaPF.PowerSystem.bounds — Function

bounds(pf::AbstractPowerSystem, attr::AbstractNetworkAttribute, val::AbstractNetworkValues)

Return lower and upper bounds corresponding to the admissible values of the AbstractNetworkAttribute attr.

Examples

p_min, p_max = bounds(pf, Generator(), ActivePower())
v_min, v_max = bounds(pf, Buses(), VoltageMagnitude())